Adicionado SeedingService

+++ SalesWeb/Models/SalesRecord.cs
@@ -7,7 +7,7 @@ namespace SalesWeb.Models
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public int Amount { get; set; }
+        public double Amount { get; set; }
         public SaleStatus Status { get; set; }
         public Seller Seller { get; set; }
 
@@ -15,7 +15,7 @@ namespace SalesWeb.Models
         {
         }
 
-        public SalesRecord(int id, DateTime date, int amount, SaleStatus status, Seller seller)
+        public SalesRecord(int id, DateTime date, double amount, SaleStatus status, Seller seller)
         {
             Id = id;
             Date = date;
